Updated with working child_ideal prefetcher

+++ ChampSim/inc/memory_class.h
@@ -50,7 +50,7 @@ public:
 
   uint8_t asid[2] = {std::numeric_limits<uint8_t>::max(), std::numeric_limits<uint8_t>::max()}, type = 0;
 
-  uint32_t pf_metadata = 0;
+  uint64_t pf_metadata = 0;
   uint32_t cpu = std::numeric_limits<uint32_t>::max();
 
   uint64_t address = 0, v_address = 0, data = 0, instr_id = 0, ip = 0, event_cycle = std::numeric_limits<uint64_t>::max(), cycle_enqueued = 0;
