+++ ChampSim/prefetcher/ip_stride/ip_stride.cc
@@ -91,13 +91,13 @@ void CACHE::prefetcher_initialize() { std::cout << NAME << " IP-based stride pre
 
 void CACHE::prefetcher_cycle_operate() { ::trackers[this].advance_lookahead(this); }
 
-uint32_t CACHE::prefetcher_cache_operate(uint64_t addr, uint64_t ip, uint8_t cache_hit, uint8_t type, uint32_t metadata_in)
+uint64_t CACHE::prefetcher_cache_operate(uint64_t addr, uint64_t ip, uint8_t cache_hit, uint8_t type, uint64_t metadata_in)
 {
   ::trackers[this].initiate_lookahead(ip, addr >> LOG2_BLOCK_SIZE);
   return metadata_in;
 }
 
-uint32_t CACHE::prefetcher_cache_fill(uint64_t addr, uint32_t set, uint32_t way, uint8_t prefetch, uint64_t evicted_addr, uint32_t metadata_in)
+uint64_t CACHE::prefetcher_cache_fill(uint64_t addr, uint32_t set, uint32_t way, uint8_t prefetch, uint64_t evicted_addr, uint64_t metadata_in)
 {
   return metadata_in;
 }
